requirements.txt
numpy
tensorflow
opencv-python
scipy
pillow
h5py